+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- macOS: wipe Quick Look cache during `sync`
+
 ## [0.15.1] - 2018-06-10
 
 ### Fixed